+++ public/base/js/util.js
@@ -1,0 +1,7 @@
+  class Util {
+    static sleep(ms) {
+      return new Promise((r) => setTimeout(r, ms));
+    }
+  }
+
+  export { Util };
