+++ public/components/header/product-search-form/scss/product-search-form.scss
@@ -1,4 +1,5 @@
 .product-search-form {
+  position: relative;
   display: flex;
   align-items: center;
   margin: 15px 0 0 20px;
