+++ public/components/header/top-bar/top-bar.scss
