+++ public/index.js
@@ -1,18 +1,20 @@
-import TopBar from './components/top-bar/topBar.js';
-import Gnb from './components/gnb/gnb.js';
-import ProductSearchForm from './components/product-search-form/productSearchForm.js';
-import RenewalHeader from './layout/renewal-header/renewalHeader.js';
+import TopBar from './components/header/top-bar/topBar.js';
+import Gnb from './components/header/gnb/gnb.js';
+import { ProductSearchForm } from './components/header/product-search-form/js/productSearchForm.js';
+import { AutoTextMaker } from './components/header/product-search-form/js/autoTextMaker.js';
 
-class Controller {
+import Header from './components/header/Header.js';
+
+class App {
   constructor() {
     this.render();
+    new AutoTextMaker();
   }
 
   render() {
-    const renewalHeader = new RenewalHeader();
-    document.querySelector('#root').appendChild(renewalHeader);
-    
+    const header = new Header();
+    document.querySelector('#root').appendChild(header);
   }
 }
 
-const controller = new Controller();
+new App();
